requirements.txt
Django==1.9.6
django-braces==1.8.1
django-extensions==1.6.7
django-filter==0.13.0
django-oauth-toolkit==0.10.0
django-rest-framework-social-oauth2==1.0.4
django-shell-plus==1.1.7
djangorestframework==3.3.3
djangorestframework-oauth==1.1.0
flufl.enum==4.1
Markdown==2.6.6
oauthlib==1.0.3
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.19
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
Werkzeug==0.11.10
